Add find_all_by_package.
Due to splitting of indexes for ES6, we can no longer rely on parent,
child relations to determine 'which packages have which versions'
because ES6 cannot join between two indexes.

Instead we lookup in the versions table a particular package (CP).
This should yield some reasonable count of versions (typically 1-100).
We then use those results to find answers like "highest version" by
simplying sorting the result and taking the first of the sort.