Fix sidekiq initializer.
The previous iteration was incorrect and probably broken. Now two
changes are coming:

Make redis configurable. It defaults to localhost:default_redis_port
but can be specified in REDIS_URL to point somewhere else. We use
this in the docker model to point it at a redis container.

Fix the initializer, I think it needs to require sidekiq to work such
that the logs are squelched.