Gemfile.lock
Consider bundling with dev-ruby/bundler 1.17.3 [not stable via gentoo system] or consider bundling with dev-ruby/bundler 1.16.1, or perhaps move away from bundler and go directly to using dev-lang/ruby and use the current bundler version 2.0.1. I am doing a pull request on a different branch to see how this works, with using bundler, and see what gems will be required to be dowgraded. It will probably be the same as the branch 25.