Migrate to ES 7.3 and the repository pattern
Elasticsearch-persistence is used as the persistence layer for Ruby
domain objects in Elasticsearch in this application. So far, the
ActiveRecord pattern has been used here. However, this pattern
has been deprecated as of version 6 of the gem and was removed in
version 7. That's why the application has been migrated to use the
repository pattern instead.

For further information, please see:
https://www.elastic.co/blog/activerecord-to-repository-changing-
persistence-patterns-with-the-elasticsearch-rails-gem

Note: The old Elasticsearch index won't be compatible with this
version anymore. That's why a fresh index should be populated.

Signed-off-by: Max Magorsch <max@magorsch.de>