Add an ES index containing the git commits
So far, any history-related information has been fetched using git.
That is, whenever a user requested any history-related information
'git log' was run and the output was parsed. That is time-consuming.
Loading the page https://packages.gentoo.org/packages/keyworded takes
around 120 seconds this way.

Instead of doing so, any git commits are now added to an ES index and
retrieved using ES. This way, the same page as mentioned before, loads
in under 3 seconds.

The commits for populating the index are fetched incrementally. This
way, the first run may take some time, but afterwards, updates are fast.

Signed-off-by: Max Magorsch <max@magorsch.de>