Keep the footer at the bottom of the page
Whenever the content of the page is taller than the viewport, the
footer is currently kept correctly beneath the content. If, however,
the content is too small to fill the viewport, the footer is displayed
in the middle of the page. For instance, please note:
 - https://packages.gentoo.org/about
 - https://packages.gentoo.org/about/help
 - https://packages.gentoo.org/arches

The introduced change keeps the footer at the bottom of the page
no matter how tall the content is.

Signed-off-by: Max Magorsch <max@magorsch.de>